pyproject.toml
[project]
name = "enable-your-vibe"
version = "1.0.0"
description = "Vibe Coding 핸즈온 세션 교육 자료"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "fastmcp>=2.0.0",
    "httpx>=0.28.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.format]
quote-style = "double"
